Done: UISearchBar Interface as well as API calls. TODO: 2nd end point call, I have a good idea for it, I am going to do it alongside the cleaning up of my SOQuery (going to probably work on making it more portable) it seems weird being a sole method within a object... I think I should either make it global or implement it into an extension, but both seem weird to me as well.
I also will use NSUserDefaults to see if this is the first time a user has opened up a page. Furthermore I will design a UIWebView to be used to open up a question when a tableViewCell is clicked.
I think I'll have time to do the challenge. It seems weird, but YOLO.